requirements.txt
#Flask==2.0.0
Werkzeug==2.0.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.0
alembic==1.4.3
astroid==2.4.2
click==7.1.2
colorama==0.4.3
dnspython==2.0.0
email-validator==1.1.1
Flask-Migrate==2.5.3

Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
isort==5.5.3
itsdangerous==1.1.0

lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.6.0
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0

toml==0.10.1

wrapt==1.12.1
WTForms==2.3.3
